Alarms Restyling PT1
The behavior of the alarms has been totally changed. The previous logic was fundamentally flawed and disinclined to change. Part of the code has been rewritten (NewAlarmFragment) to allow the scheduling of multiple alarms and their dismissal.
Although this commit is considered stable enough to test the application, no type of functional testing has been carried out atm. It is therefore still not considered usable on a daily basis.